Add transalation when creating contact and update contsants location
