+++ app/javascript/src/translations/en.json
@@ -38,6 +38,7 @@
       "email": "Email",
       "createdAt": "Created At"
     },
+    "createSuccess": "Contact created successfully",
     "fields": {
       "firstName": "First Name",
       "lastName": "LastName",
